made sample time an u64

+++ src/time.rs
@@ -4,4 +4,4 @@ pub use beats::{BeatTimeBase, BeatTimeStep};
 // -------------------------------------------------------------------------------------------------
 
 /// Sample time value type emitted by the [Pattern](`crate::Pattern`) trait.
-pub type SampleTime = usize;
+pub type SampleTime = u64;
